Avoid waiting for crashed sonic docker
